pyproject.toml
[project]
name = "vercel-chatbot-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "uvicorn>=0.38.0",
    "vercel-ai>=0.1.0",
]
